+++ build-scripts/builder/entrypoint.sh
@@ -5,12 +5,14 @@ set -e
 
 # Set default values for environment variables
 INDEXER_PLUGINS_BRANCH=${INDEXER_PLUGINS_BRANCH:-main}
+INDEXER_REPORTING_BRANCH=${INDEXER_REPORTING_BRANCH:-main}
 REVISION=${REVISION:-0}
 IS_STAGE=${IS_STAGE:-false}
 DISTRIBUTION=${DISTRIBUTION:-rpm}
 ARCHITECTURE=${ARCHITECTURE:-x64}
 
 PLUGINS_REPO_DIR="/repositories/wazuh-indexer-plugins"
+REPORTING_REPO_DIR="/repositories/wazuh-indexer-reporting"
 
 # Function to clone repositories
 clone_repositories() {
@@ -22,6 +24,12 @@ clone_repositories() {
         git -C "$PLUGINS_REPO_DIR" checkout "$INDEXER_PLUGINS_BRANCH"
     else
         git clone --branch "$INDEXER_PLUGINS_BRANCH" https://github.com/wazuh/wazuh-indexer-plugins --depth 1 "$PLUGINS_REPO_DIR"
+    fi
+
+    if [ -d "$REPORTING_REPO_DIR/.git" ]; then
+        git -C "$REPORTING_REPO_DIR" checkout "$INDEXER_REPORTING_BRANCH"
+    else
+        git clone --branch "$INDEXER_REPORTING_BRANCH" https://github.com/wazuh/wazuh-indexer-reporting --depth 1 "$REPORTING_REPO_DIR"
     fi
 }
 
@@ -37,6 +45,18 @@ build_plugins() {
     ./gradlew build -Dversion="$version" -Drevision="$revision" --no-daemon
 }
 
+# Function to build wazuh-indexer-reporting
+build_reporting() {
+    echo "----------------------------------------"
+    echo "Building Reporting"
+    echo "----------------------------------------"
+    local version="$1"
+    local revision="$2"
+    cd ${REPORTING_REPO_DIR}
+    echo "Building reporting..."
+    ./gradlew build -Dversion="$version" -Drevision="$revision" --no-daemon
+}
+
 # Function to copy builds
 copy_builds() {
     echo "----------------------------------------"
@@ -47,6 +67,8 @@ copy_builds() {
     mkdir -p ~/artifacts/plugins
     echo "Copying setup plugin..."
     cp ${PLUGINS_REPO_DIR}/plugins/setup/build/distributions/wazuh-indexer-setup-"$version"."$revision".zip ~/artifacts/plugins
+    echo "Copying reporting..."
+    cp ${REPORTING_REPO_DIR}/build/distributions/wazuh-indexer-reports-scheduler-"$version"."$revision".zip ~/artifacts/plugins
 }
 
 # Function for packaging process
@@ -60,10 +82,13 @@ package_artifacts() {
     local is_stage="$4"
 
     local plugins_hash
+    local reporting_hash
     local package_min_name
     local package_name
 
     plugins_hash=$(cd ${PLUGINS_REPO_DIR} && git rev-parse --short HEAD)
+
+    reporting_hash=$(cd ${REPORTING_REPO_DIR} && git rev-parse --short HEAD)
 
     cd ~
 
@@ -73,6 +98,7 @@ package_artifacts() {
         -d "$distribution" \
         -r "$revision" \
         -l "$plugins_hash" \
+        -e "$reporting_hash" \
         "$(if [ "$is_stage" = "true" ]; then echo "-x"; fi)")
 
     echo "Creating package name..."
@@ -81,6 +107,7 @@ package_artifacts() {
         -d "$distribution" \
         -r "$revision" \
         -l "$plugins_hash" \
+        -e "$reporting_hash" \
         "$(if [ "$is_stage" = "true" ]; then echo "-x"; fi)")
 
     echo "Building package..."
@@ -91,6 +118,7 @@ package_artifacts() {
         -d "$distribution" \
         -r "$revision" \
         -l "$plugins_hash" \
+        -e "$reporting_hash" \
 
 }
 
@@ -102,6 +130,7 @@ main() {
     VERSION=$(bash ~/build-scripts/product_version.sh)
     # Build and assemble the package
     build_plugins "$VERSION" "$REVISION"
+    build_reporting "$VERSION" "$REVISION"
     copy_builds "$VERSION" "$REVISION"
     package_artifacts "$ARCHITECTURE" "$DISTRIBUTION" "$REVISION" "$IS_STAGE"